+++ app/src/main/AndroidManifest.xml
@@ -10,28 +10,28 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Quotes">
         <activity
-            android:name=".ForgottenPassword"
+            android:name=".activity.ForgottenPassword"
             android:exported="false"
             android:label=""/>
         <activity
-            android:name=".UserProfile"
+            android:name=".activity.UserProfile"
             android:exported="false"
             android:label=""/>
         <activity
-            android:name=".RegisterActivity"
+            android:name=".activity.RegisterActivity"
             android:exported="false"
             android:label=""/>
         <activity
-            android:name=".LoginActivity"
+            android:name=".activity.LoginActivity"
             android:exported="false"
             android:label=""/>
         <activity
-            android:name=".CreateQuotes"
+            android:name=".activity.CreateQuotes"
             android:exported="false"
             android:label="Create A Quote"
-            android:parentActivityName=".MainActivity"/>
+            android:parentActivityName=".activity.MainActivity"/>
         <activity
-            android:name=".MainActivity"
+            android:name=".activity.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.AppCompat.SlidrActivityTheme">
             <intent-filter>
