+++ app/src/main/java/com/sforge/quotes/activity/ForgottenPassword.java
@@ -1,4 +1,4 @@
-package com.sforge.quotes;
+package com.sforge.quotes.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -12,6 +12,7 @@ import android.widget.Toast;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.google.firebase.auth.FirebaseAuth;
+import com.sforge.quotes.R;
 
 public class ForgottenPassword extends AppCompatActivity {
