+++ app/src/main/java/com/sforge/quotes/activity/RegisterActivity.java
@@ -1,4 +1,4 @@
-package com.sforge.quotes;
+package com.sforge.quotes.activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -13,8 +13,9 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.google.firebase.auth.FirebaseAuth;
-import com.sforge.quotes.QuoteEntity.DAOQuote;
-import com.sforge.quotes.UserEntity.User;
+import com.sforge.quotes.R;
+import com.sforge.quotes.repository.DAOQuote;
+import com.sforge.quotes.entity.User;
 
 import java.util.Objects;
