+++ app/src/main/java/com/sforge/quotes/entity/Quote.java
@@ -1,4 +1,4 @@
-package com.sforge.quotes.QuoteEntity;
+package com.sforge.quotes.entity;
 
 public class Quote {
