+++ app/src/main/java/com/sforge/quotes/entity/User.java
@@ -1,4 +1,4 @@
-package com.sforge.quotes.UserEntity;
+package com.sforge.quotes.entity;
 
 public class User {
