+++ app/src/main/java/com/sforge/quotes/repository/DAOQuote.java
@@ -1,9 +1,10 @@
-package com.sforge.quotes.QuoteEntity;
+package com.sforge.quotes.repository;
 
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.Query;
+import com.sforge.quotes.entity.Quote;
 
 import java.util.HashMap;
