+++ app/src/main/res/layout/activity_create_quotes.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CreateQuotes"
+    tools:context=".activity.CreateQuotes"
     android:orientation="vertical"
     android:padding="8dp">
