+++ app/src/main/res/layout/activity_forgotten_password.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ForgottenPassword">
+    tools:context=".activity.ForgottenPassword">
 
     <TextView
         android:id="@+id/forgotPasswordTitle"
