+++ app/src/main/res/layout/activity_user_profile.xml
@@ -5,7 +5,7 @@
     android:background="@color/transparent"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UserProfile">
+    tools:context=".activity.UserProfile">
 
     <TextView
         android:id="@+id/profileUsername"
